Convert to using size_t instead of int.

+++ src/libsrcml/srcml.h
@@ -208,7 +208,7 @@ LIBSRCML_DECL int srcml_set_hash                  (const char* hash);
 LIBSRCML_DECL int srcml_set_options               (unsigned long long option);
 LIBSRCML_DECL int srcml_enable_option             (unsigned long long option);
 LIBSRCML_DECL int srcml_disable_option            (unsigned long long option);
-LIBSRCML_DECL int srcml_set_tabstop               (int tabstop);
+LIBSRCML_DECL int srcml_set_tabstop               (size_t tabstop);
 LIBSRCML_DECL int srcml_register_file_extension   (const char* extension, const char* language);
 LIBSRCML_DECL int srcml_register_namespace        (const char* prefix, const char* ns);
 LIBSRCML_DECL int srcml_set_processing_instruction(const char* target, const char* data); 
@@ -224,7 +224,7 @@ LIBSRCML_DECL const char*        srcml_get_version  ();
 LIBSRCML_DECL const char*        srcml_get_timestamp();
 LIBSRCML_DECL const char*        srcml_get_hash();
 LIBSRCML_DECL unsigned long long srcml_get_options  ();
-LIBSRCML_DECL int                srcml_get_tabstop  ();
+LIBSRCML_DECL size_t             srcml_get_tabstop  ();
 LIBSRCML_DECL const char*        srcml_get_processing_instruction_target(); 
 LIBSRCML_DECL const char*        srcml_get_processing_instruction_data  (); 
 
@@ -233,31 +233,31 @@ LIBSRCML_DECL const char*        srcml_get_processing_instruction_data  ();
 */
 
 /* Number of declared XML namespaces */
-LIBSRCML_DECL int                srcml_get_namespace_size();
+LIBSRCML_DECL size_t             srcml_get_namespace_size();
 
 /* Prefix of the namespace at that position, where empty namespace is an empty string, and 0 is invalid position */
-LIBSRCML_DECL const char*        srcml_get_namespace_prefix(int pos);
+LIBSRCML_DECL const char*        srcml_get_namespace_prefix(size_t pos);
 
 /* Prefix of the namespace with this namespace uri */
 LIBSRCML_DECL const char*        srcml_get_prefix_from_uri(const char* namespace_uri);
 
 /* URI of the namespace at that position, where 0 is invalid position */
-LIBSRCML_DECL const char*        srcml_get_namespace_uri(int pos);
+LIBSRCML_DECL const char*        srcml_get_namespace_uri(size_t pos);
 
 /* URI of the namespace with this namespace prefix */
 LIBSRCML_DECL const char*        srcml_get_uri_from_prefix(const char* prefix);
 
-LIBSRCML_DECL int                srcml_get_macro_list_size();
-LIBSRCML_DECL const char*        srcml_get_macro_token(int pos);
+LIBSRCML_DECL size_t             srcml_get_macro_list_size();
+LIBSRCML_DECL const char*        srcml_get_macro_token(size_t pos);
 LIBSRCML_DECL const char*        srcml_get_macro_token_type(const char* namespace_uri);
-LIBSRCML_DECL const char*        srcml_get_macro_type(int pos);
+LIBSRCML_DECL const char*        srcml_get_macro_type(size_t pos);
 
 /* Source-code language is supported */
 LIBSRCML_DECL int srcml_check_language(const char* language);
 
 /* Query supported source-code languages */
-LIBSRCML_DECL int srcml_get_language_list_size();
-LIBSRCML_DECL const char* srcml_get_language_list(int pos);
+LIBSRCML_DECL size_t srcml_get_language_list_size();
+LIBSRCML_DECL const char* srcml_get_language_list(size_t pos);
 
 /* Currently registered language for a file extension
    When full filename is given, the extension is extracted */
@@ -291,7 +291,7 @@ LIBSRCML_DECL struct srcml_archive* srcml_archive_clone(const struct srcml_archi
 
 /* Open a srcML archive for output */
 LIBSRCML_DECL int srcml_archive_write_open_filename(struct srcml_archive*, const char* srcml_filename);
-LIBSRCML_DECL int srcml_archive_write_open_memory  (struct srcml_archive*, char** buffer, int * size);
+LIBSRCML_DECL int srcml_archive_write_open_memory  (struct srcml_archive*, char** buffer, size_t * size);
 LIBSRCML_DECL int srcml_archive_write_open_FILE    (struct srcml_archive*, FILE* srcml_file);
 LIBSRCML_DECL int srcml_archive_write_open_fd      (struct srcml_archive*, int srcml_fd);
 LIBSRCML_DECL int srcml_archive_write_open_io      (struct srcml_archive*, void * context, int (*write_callback)(void * context, const char * buffer, int len), int (*close_callback)(void * context));
@@ -307,7 +307,7 @@ LIBSRCML_DECL int srcml_archive_set_version            (struct srcml_archive*, c
 LIBSRCML_DECL int srcml_archive_set_options               (struct srcml_archive*, unsigned long long option);
 LIBSRCML_DECL int srcml_archive_enable_option             (struct srcml_archive*, unsigned long long option);
 LIBSRCML_DECL int srcml_archive_disable_option            (struct srcml_archive*, unsigned long long option);
-LIBSRCML_DECL int srcml_archive_set_tabstop               (struct srcml_archive*, int tabstop);
+LIBSRCML_DECL int srcml_archive_set_tabstop               (struct srcml_archive*, size_t tabstop);
 LIBSRCML_DECL int srcml_archive_register_file_extension   (struct srcml_archive*, const char* extension, const char* language);
 LIBSRCML_DECL int srcml_archive_register_namespace        (struct srcml_archive*, const char* prefix, const char* ns);
 LIBSRCML_DECL int srcml_archive_set_processing_instruction(struct srcml_archive*, const char* target, const char* data); 
@@ -322,18 +322,18 @@ LIBSRCML_DECL const char*        srcml_archive_get_filename                     
 LIBSRCML_DECL const char*        srcml_archive_get_directory                    (const struct srcml_archive*);
 LIBSRCML_DECL const char*        srcml_archive_get_version                      (const struct srcml_archive*);
 LIBSRCML_DECL unsigned long long srcml_archive_get_options                      (const struct srcml_archive*);
-LIBSRCML_DECL int                srcml_archive_get_tabstop                      (const struct srcml_archive*);
-LIBSRCML_DECL int                srcml_archive_get_namespace_size               (const struct srcml_archive*);
-LIBSRCML_DECL const char*        srcml_archive_get_namespace_prefix             (const struct srcml_archive*, int pos);
+LIBSRCML_DECL size_t             srcml_archive_get_tabstop                      (const struct srcml_archive*);
+LIBSRCML_DECL size_t             srcml_archive_get_namespace_size               (const struct srcml_archive*);
+LIBSRCML_DECL const char*        srcml_archive_get_namespace_prefix             (const struct srcml_archive*, size_t pos);
 LIBSRCML_DECL const char*        srcml_archive_get_prefix_from_uri              (const struct srcml_archive*, const char* namespace_uri);
-LIBSRCML_DECL const char*        srcml_archive_get_namespace_uri                (const struct srcml_archive*, int pos);
+LIBSRCML_DECL const char*        srcml_archive_get_namespace_uri                (const struct srcml_archive*, size_t pos);
 LIBSRCML_DECL const char*        srcml_archive_get_uri_from_prefix              (const struct srcml_archive*, const char* prefix);
 LIBSRCML_DECL const char*        srcml_archive_get_processing_instruction_target(const struct srcml_archive*); 
 LIBSRCML_DECL const char*        srcml_archive_get_processing_instruction_data  (const struct srcml_archive*); 
-LIBSRCML_DECL int                srcml_archive_get_macro_list_size              (const struct srcml_archive*);
-LIBSRCML_DECL const char*        srcml_archive_get_macro_token                  (const struct srcml_archive*, int pos);
+LIBSRCML_DECL size_t             srcml_archive_get_macro_list_size              (const struct srcml_archive*);
+LIBSRCML_DECL const char*        srcml_archive_get_macro_token                  (const struct srcml_archive*, size_t pos);
 LIBSRCML_DECL const char*        srcml_archive_get_macro_token_type             (const struct srcml_archive*, const char* namespace_uri);
-LIBSRCML_DECL const char*        srcml_archive_get_macro_type                   (const struct srcml_archive*, int pos);
+LIBSRCML_DECL const char*        srcml_archive_get_macro_type                   (const struct srcml_archive*, size_t pos);
 
 /* Create a new srcml unit.
    Client is responsible for freeing memory using srcml_unit_free() */
@@ -404,7 +404,7 @@ LIBSRCML_DECL const char* srcml_unit_get_formatted_xml(struct srcml_unit*, const
 
 /* Convert from srcML to source code */
 LIBSRCML_DECL int srcml_unit_unparse_filename(struct srcml_unit*, const char* src_filename);
-LIBSRCML_DECL int srcml_unit_unparse_memory  (struct srcml_unit*, char** src_buffer, int * src_size);
+LIBSRCML_DECL int srcml_unit_unparse_memory  (struct srcml_unit*, char** src_buffer, size_t * src_size);
 LIBSRCML_DECL int srcml_unit_unparse_FILE    (struct srcml_unit*, FILE* srcml_file);
 LIBSRCML_DECL int srcml_unit_unparse_fd      (struct srcml_unit*, int srcml_fd);
 LIBSRCML_DECL int srcml_unit_unparse_io      (struct srcml_unit*, void * context, int (*write_callback)(void * context, const char * buffer, int len), int (*close_callback)(void * context));
